Move queries and types to their own files
