+++ crates/sui-graphql-client/README.md
@@ -13,7 +13,7 @@ executing transactions and more.
 # Usage
 
 ## Connecting to a GraphQL server
-Instantiate [`Client`] with [`Client::new(server: &str)`] or use one of the predefined functions for different networks [`Client`].
+Instantiate a client with [`Client::new(server: &str)`] or use one of the predefined functions for different networks [`Client`].
 
 ```rust
 use sui_graphql_client::Client;
@@ -22,7 +22,7 @@ use anyhow::Result;
 #[tokio::main]
 async fn main() -> Result<()> {
 
-   // Connect to default testnet GraphQL server
+   // Connect to the mainnet GraphQL server
    let client = Client::new_mainnet()?;
    let chain_id = client.chain_id().await?;
    println!("{:?}", chain_id);
@@ -92,8 +92,10 @@ The complete example is shown below:
 use anyhow::Result;
 use cynic::QueryBuilder;
 
-use sui_graphql_client::graphql_types::{schema, BigInt, Uint53, ObjectFilter};
-use sui_graphql_client::Client;
+use sui_graphql_client::{
+    query_types::{schema, BigInt, Uint53},
+    Client,
+};
 use sui_types::types::Address;
 
 // The data returned by the custom query.
@@ -114,6 +116,7 @@ pub struct EpochData {
 pub struct CustomVariables {
     pub id: Option<Uint53>,
 }
+
 // The custom query. Note that the variables need to be explicitly declared.
 #[derive(cynic::QueryFragment, Debug)]
 #[cynic(schema = "rpc", graphql_type = "Query", variables = "CustomVariables")]
